requirements.txt
## The following requirements were added by pip freeze:
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-material==1.11.3
django-material-admin==1.8.6
django-phonenumber-field==7.0.0
phonenumberslite==8.12.56
Pillow==9.2.0
six==1.16.0
sqlparse==0.4.2